Revert "test(bench): prevent compiler optimisations"
This reverts commit 867d98627ac0cbcae6f4e00053dca7d278f7714c.

This feature requires nightly. This is only a test dependency, so it
would actually be fine if this feature was only enabled when running the
tests on a nightly compiler, but so far I have not found a way to do that
easily. Once we do, we can revert the revert, and use this feature on
nightly test runs to get better benchmarking insights.