chore: add cargo-readme as dev dependency
This will be integrated into the release pipeline in the future, but for
now you can execute the following command to keep the README and library
documentation in-sync:

    $ cargo readme --no-indent-headings > README.md

Note that when running this command, there will be some unintended changes
that should be discarded. Whenever I find the time, I will investigate and
try to fix the issues upstream.