feat: rewrite as_enum to support mutable machines
Before, you'd write the following:

    let sm = Machine::new(MyState);
    match sm.state().as_enum() {
        States::MyState => {
            // ...
        },
        States::OtherState => {
            // ...
        }
    }

The downside of this was that you could not wrap this in a block, for
example this would not compile:

    let mut sm = Machine::new(MyState);

    loop {
        sm = match sm.state().as_enum() {
            States::MyState => {
                sm.event(MyEvent)
            },
            States::OtherState => {}
        }
    }

It would fail, because the initial type of `sm` is `Machine<MyState>`, but
the transition triggered by `MyEvent` would (potentially) return a
different type, such as `Machine<OtherState>`, which means you can't
mutate the original `sm` into this new value, since it has a different
type signature.

With this change, the enum variants are no longer of type `State`, but of
type `Machine<State>`. This means you can now do the following:

    let mut sm = Machine::new(MyState);

    loop {
        sm = match sm.state().as_enum() {
            States::MyState => {
                sm.event(MyEvent)
            },
            States::OtherState => {}
        }
    }

You can now do the following:

    let mut sm = Machine::new(MyState).as_enum();

    loop {
        sm = match sm {
            States::MyState(m) => {
                m.event(MyEvent).as_enum()
            },
            States::OtherState => {}
        }
    }

In this case, you match over `Machine<State>` variants, and each variant
is a tuple value with the machine returned as `m`. You can then transition
the machine, and (optionally) convert it back to an enum using
`as_enum()`.

In terms of performance, this change has no impact over the previous enum
implementation. There is no significant impact when you do a single match:

    GameLoop; Idle => Simulating => Idle => Rendering => Idle
                            time:   [269.58 ps 271.54 ps 273.82 ps]

    GameLoop – match; Idle => Simulating => Idle => Rendering => Idle
                            time:   [290.13 ps 294.21 ps 298.76 ps]

But performance does decrease if you switch to a loop construct:

    GameLoop – loop; Idle => Simulating => Idle => Rendering => Idle
                            time:   [1.5244 ns 1.5422 ns 1.5596 ns]

Keep this in mind when designing your flow of logic.

BREAKING CHANGE:

The old `sm.state().as_enum()` has been replaced with `sm.as_enum()`.