refactor: introduce Machine trait
This trait currently has a single method signature to query the state of
the state machine.

    fn state(&self) -> Self::State

This change is another step towards allowing usage of the crate without
the need for the macro, and to genericise as much implementation details
as possible, to only require a small subset of concrete type
implementations to be defined for a working state machine.