refactor: remove concrete type from AsEnum trait
This allows us to more easily move the `AsEnum` trait out of the macro.

This change is another step towards allowing usage of the crate without
the need for the macro, and to genericise as much implementation details
as possible, to only require a small subset of concrete type
implementations to be defined for a working state machine.