refactor: add top-level workspace Cargo.toml
This is in preparation of migrating the macro_rules macro to the new
proc-macro implementation, which has to be implemented in its own crate,
separate from the exposed traits in the current crate.

With this workspace Cargo.toml, the repository now supports multiple
crates that can still be built/tested using a top-level `cargo test`
command.