test: introduce new "test_suite" integration tests
This is a separate crate from the "sm" and "sm_macro" crates. It's purpose
is to test the two sm crates together and verify their correctness.

This idea is taken from the "serde" crate, which has a similar structure.
It works well, except that you need to do some shenanigans to set up the
suite first, and then run the tests. For now this is okay, we can
experiment with different setups later.

The old tests are kept as-is for now, but will probably be migrated to one
single test suite, or be removed entirely if they are duplicates.