chore: remove cargo-release integration
cargo-release does not support workspace-wide releases (meaning, all
crates in the workspace should be updated to the same version in the
same commit).

There is ongoing work to make this possible:

* https://github.com/spacekookie/cargo-ws-release
* https://github.com/sunng87/cargo-release/issues/66

For the next couple of releases, I will follow in the footsteps of the
previous "cargo-release"-releases, and hopefully either workspace support
becomes available by then, or I find some spare cycles to help finish the
work that was started on that feature.