feat: make sm_macro a feature of the sm crate
With this change, the "sm" crate is the only crate you interact with as a
user.

When you want to use the sm macro, you write:

    extern crate sm;
    use sm::sm;

If you leave off the `use sm::sm;` part, the macro will not be imported
into your scope, but will still be compiled into your binary.

The macro functionality is provided by the by-default enabled "macro"
feature in the sm crate.

If you only care about the functionality provided by the sm crate
(meaning, the public traits), you can use the following in your Cargo.toml
definition:

    [dependencies]
    sm = { version = "*", default-features = false }