feat: restrict machines to a fixed set of initial states
With this change, the no-op `States {}` declaration is
replaced with the `InitialStates {}` declaration.

Any states defined in the `InitialStates` block can be
used as the initial state of the state machine:

    let sm = TurnStile::Machine::new(TurnStile::Locked);

Any other state part of the state machine but not part
of the list of initial states will result in a compile
time error when used during initialization.

BREAKING CHANGE:

This change instantly removes the `States` declaration
(it was no longer doing anything since the upgrade to
procedural macro's, but was still around). The
compiler will return an error when using `States`
instead of `InitialStates`.