feat: provide access to event that triggered machine state
You can now ask the machine what event caused it to end up in the current
state.

This is relevant information if you have two or more events leading to the
same state. By querying the machine for the event that triggered the state
change, you can determine how your business logic should behave.

Example:

    let sm = Machine::new(Locked);
    assert!(sm.trigger().is_none());

    let sm = sm.transition(TurnKey);
    assert_eq!(sm.trigger.unwrap(), TurnKey);

`Machine` implements a `trigger()` method that returns an optional event.
The return value is `None` if the machine is still in its initial state
before any state transitions have been executed. After the first
transition, the return value will always be a valid event that was used to
trigger the last transition.

BREAKING CHANGE:

The enum now has a generic type `E: Event` attached to it. This makes it
impossible to have a mutating variable that holds the `States` enum, and
replace it with a new enum after a transition.

This is a (very) useful pattern in loop-based event machine logic.

The problem is best seen in the `enum-wrapped-machine` test. This test has
been disabled for now, but will be re-enabled after this problem has been
tackled in the next release. This also means that the pattern used in that
test is unavailable for the time being.