feat: re-introduce enum-based state matching loops
This functionality was disabled since
49b58fdee676249278f4e94a711c7efdd1396a48.

The functionality is re-introduced, with support for the recently added
`trigger` implementation that allows you to tell which event triggered the
current machine state.

Usage is the same as before, but the enum variants have been altered to
support this new concept:

    use TurnStile::*;
    let sm = Machine::new(Unlocked);

    match sm.as_enum() {
        Variant::InitialUnlocked(_) => {}
        Variant::UnlockedByCoin(_) => {}
        Variant::LockedByPush(m) => {
            assert_eq!(m.state(), Locked);
            assert_eq!(m.trigger().unwrap(), Push);
        }
    }

The enum variants are named `<state>By<event>`, which helps to distinguish
between the same states, triggered by different events.

The initial state of a machine has no trigger event, and is named
`Initial<state>`. If a machine supports multiple initial states, all of
them need to be added to the exhaustive pattern match.

There is (some) runtime overhead in doing pattern matching. Benchmarking
shows ~300 ps (picoseconds) or 0.3 nanoseconds of runtime when using the
state machine without pattern matching, and 1500ps, or 1.5 nanoseconds of
runtime overhead when pattern matching the current state.